fix: refactor files and change location of some files
